add test, move caching tests closer to code
